[TS-267] Rename: categoryConstants -> starCategoryConstants로 이름 변경
