+++ src/constants/starCategoryConstants.tsx
